requirements.txt
appdirs==1.4.4
build==1.2.2.post1
buildozer==1.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Cython==0.29.33
distlib==0.3.9
docutils==0.21.2
filelock==3.16.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
Kivy==2.3.0
Kivy-Garden==0.1.5
MarkupSafe==3.0.2
packaging==24.2
pexpect==4.9.0
platformdirs==4.3.6
plyer==2.1.0
ptyprocess==0.7.0
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
requests==2.32.3
sh==1.14.3
toml==0.10.2
tomli==2.2.1
urllib3==2.2.3
virtualenv==20.28.0
zipp==3.21.0
